Split off `Customer` and key derivation into separate module
